added separate test class for json schema
